Auth refactor

+++ app/core/services/company_service.py
@@ -2,11 +2,10 @@ from dataclasses import dataclass
 from typing import Optional
 
 from app.core.constants import Status
-from app.core.models import Company, Industry, OrganizationSize
+from app.core.models import Account, Company, Industry, OrganizationSize
 from app.core.repository.company_repository import ICompanyRepository
 
 
-# TODO: RAISE CORRECT EXCEPTIONS
 @dataclass
 class CompanyService:
     company_repository: ICompanyRepository
@@ -34,13 +33,16 @@ class CompanyService:
 
     def update_company(
         self,
-        token: str,
+        account: Account,
         company_id: int,
         name: str,
         website: str,
         industry: Industry,
         organization_size: OrganizationSize,
     ) -> tuple[Status, Optional[Company]]:
+        if company_id not in account.companies:
+            return Status.COMPANY_DOES_NOT_EXIST, None
+
         company = self.company_repository.update_company(
             company_id=company_id,
             name=name,
@@ -53,7 +55,10 @@ class CompanyService:
 
         return Status.OK, company
 
-    def delete_company(self, company_id: int) -> Status:
+    def delete_company(self, account: Account, company_id: int) -> Status:
+        if company_id not in account.companies:
+            return Status.COMPANY_DOES_NOT_EXIST
+
         if self.company_repository.delete_company(company_id=company_id):
             return Status.OK
         return Status.ERROR_DELETING_COMPANY
